Propagate tags between meta and leaf puzzles (#64)
Any tag whose name starts with the word "meta" will display in the puzzle info of all puzzles in that round.
If a meta has a tag named "cares about", whose value contains a comma-separated list of tag names, a column will be added to the table of leaf puzzles in its puzzle info for each such tag name containing that puzzle's value for that tag. If the tag isn't set, a prompt to set it will be added to the leaf puzzle's puzzle info.