Add color to rounds to help distinguish them
On the blackboard, each round is assigned a color based on a hash of its id; this can be overridden with the color tag, which accepts any CSS color. In edit mode, a color picker box is provided to set the color. On the puzzle page, the same color is used for the round and a (hopefully) different color is used for each puzzle, which again can be forced with the color tag, in the hopes that they will be distinguished by backsolvers. A control is provided to turn off the extra colors.