Favorite puzzles (#104)
* Add favorites field

Each puzzle now has a list of users who have favorited it. We use a $elemMatch field projection to filter the entries down to just yourself, both for relative privacy, not to push spurious updates over the wire, and to make filtering based on whether you've favorited a puzzle faster. Unfortunately since $elemMatch counts as an inclusion, not an exclusion, we have to explicitly enumerate every field in the Puzzles collection. Hopefully I didn't miss any. Alternatively I could exclude favorites from the regular publish and add a manual publish where I add it; this is what I would have to do if I wanted to push a favorite count without pushing the whole list.

* Add favorite button to blackboard

Button alongside spreadsheet, doc, chat, etc. links that faves/unfaves the puzzle. It's hollow when not favorited and red filled when favorited.

* Pull favorite puzzles to the top

If any puzzles are favorited, adds a section to the top of the blackboard table containing all favorited puzzles.

* Add heart to puzzle info pane

Show favorite status, and allow changing it, directly from puzzle.

* Hide solved favorites when hiding solved puzzles

* Add trailing newlines

* One more missing trailing newline