Puzzle mechanics (#108)
Puzzles have a 'mechanics' field which is a list of canonical mechanic names. Users have a 'favorite_mechanics' field they can edit with a dropdown on the main blackboard. Puzzles with any mechanic matching a user's favorites are added to their "Favorite Puzzles" section. Adds a notification stream for favorite mechanics; when enabled and a favorite mechanic is added to a puzzle, a notification is fired. Clicking on the notification takes you to the puzzle.
Fixes #92 and lays the groundwork for #107.